wip: Com objects are horrid
