+++ program_info/win_install.nsi.in
@@ -134,7 +134,7 @@ VIAddVersionKey /LANG=${LANG_ENGLISH} "ProductVersion" "@Launcher_VERSION_NAME4@
   WriteRegStr ShCtx "Software\Classes\${APP_ID}" "" `${DESCRIPTION}`
   WriteRegStr ShCtx "Software\Classes\${APP_ID}\DefaultIcon" "" `${ICON}`
   ; default open verb
-  WriteRegStr ShCtx "Software\Classes\${APP_ID}\shell" "" "open"                  
+  WriteRegStr ShCtx "Software\Classes\${APP_ID}\shell" "" "open"
   WriteRegStr ShCtx "Software\Classes\${APP_ID}\shell\open" "" `${COMMANDTEXT}`
   WriteRegStr ShCtx "Software\Classes\${APP_ID}\shell\open\command" "" `${COMMAND}`
 
@@ -153,7 +153,7 @@ VIAddVersionKey /LANG=${LANG_ENGLISH} "ProductVersion" "@Launcher_VERSION_NAME4@
 
   !insertmacro APP_SETUP_Def `${DESCRIPTION}` `${ICON}` `${APP_ID}` `${APP_NAME}` `${APP_EXE}` `${COMMANDTEXT}` `${COMMAND}`
 
-  # Register "Default Programs" 
+  # Register "Default Programs"
   WriteRegStr ShCtx "Software\Classes\Applications\${APP_EXE}\Capabilities" "ApplicationDescription" `${DESCRIPTION}`
   WriteRegStr ShCtx "Software\RegisteredApplications" `${APP_NAME}` "Software\Classes\Applications\${APP_EXE}\Capabilities"
 
@@ -231,7 +231,7 @@ VIAddVersionKey /LANG=${LANG_ENGLISH} "ProductVersion" "@Launcher_VERSION_NAME4@
   ${IfNot} ${Errors}
     DeleteRegKey ShCtx "Software\Classes\${APP_ID}\DefaultIcon"
   ${EndIf}
-  
+
   # Unregister "Open With"
   DeleteRegKey ShCtx "Software\Classes\Applications\${APP_EXE}"
 
@@ -256,7 +256,7 @@ VIAddVersionKey /LANG=${LANG_ENGLISH} "ProductVersion" "@Launcher_VERSION_NAME4@
 !macro APP_TEARDOWN_DEFAULT APP_ID APP_NAME APP_EXE
 
   !insertmacro APP_TEARDOWN_Def `${APP_ID}` `${APP_NAME}` `${APP_EXE}`
-  
+
   # Unregister "Default Programs"
   DeleteRegValue ShCtx "Software\RegisteredApplications" `${APP_NAME}`
   DeleteRegKey ShCtx "Software\Classes\Applications\${APP_EXE}\Capabilities"
@@ -321,7 +321,7 @@ Function RunUninstall
 	  GetFullPathName $3 "$2\.." ; $InstDir
 	  IfFileExists "$2" 0 +4
 	  ExecWait $4 $1 ; The file exists, call the saved command
-    IntCmp $1 0 "" +2 +2 ; Don't delete the installer if it was aborted ; 
+    IntCmp $1 0 "" +2 +2 ; Don't delete the installer if it was aborted ;
 	  Delete "$2" ; Delete the uninstaller
 	  RMDir "$3" ; Try to delete $InstDir
   Pop $4
@@ -337,7 +337,7 @@ Section "" UninstallPrevious
   ${If} $0 == ""
     ReadRegStr $0 HKCU "${UNINST_KEY}" "UninstallString"
   ${EndIf}
-  
+
   ${If} $0 != ""
 	  !insertmacro RunUninstall $0 $0
 	  ${If} $0 <> 0
@@ -389,13 +389,13 @@ Section "@Launcher_DisplayName@"
   ; Write the installation path into the registry
   WriteRegStr HKCU Software\@Launcher_CommonName@ "InstallDir" "$INSTDIR"
 
-  ; Write the URL Handler into registry for curseforge
+  ; Write the URL Handler into registry for curseforge://
   WriteRegStr HKCU Software\Classes\curseforge "URL Protocol" ""
   WriteRegStr HKCU Software\Classes\curseforge\shell\open\command "" '"$INSTDIR\@Launcher_APP_BINARY_NAME@.exe" "%1"'
 
-; Write the URL Handler into registry for prismlauncher
-  WriteRegStr HKCU Software\Classes\prismlauncher "URL Protocol" ""
-  WriteRegStr HKCU Software\Classes\prismlauncher\shell\open\command "" '"$INSTDIR\@Launcher_APP_BINARY_NAME@.exe" "%1"'
+  ; Write the URL Handler into registry for prismlauncher://
+  WriteRegStr HKCU "Software\Classes\@Launcher_APP_BINARY_NAME@" "URL Protocol" ""
+  WriteRegStr HKCU "Software\Classes\@Launcher_APP_BINARY_NAME@\shell\open\command" "" '"$INSTDIR\@Launcher_APP_BINARY_NAME@.exe" "%1"'
 
   ; Write the uninstall keys for Windows
   ; https://learn.microsoft.com/en-us/windows/win32/msi/uninstall-registry-key
@@ -447,7 +447,7 @@ SectionEnd
 Section /o "Shell Association (Open-With dialog)" SHELL_ASSOC
 
   !insertmacro APP_SETUP `${APPDESCRIPTION}` `${APPICON}` `${APPID}` `${APPCMDTEXT}` `${APPEXE}` `${APPCMDTEXT}` '$INSTDIR\${APPEXE} -I "%1"'
-  
+
   !insertmacro APP_ASSOCIATE_DEFAULT ".mrpack" `${APPID}` `${APPEXE}` true
   !insertmacro APP_ASSOCIATE ".zip" `${APPID}` `${APPEXE}` false
 
@@ -489,7 +489,7 @@ Section "Uninstall"
 SectionEnd
 
 Section -un.ShellAssoc
-  
+
   !insertmacro APP_TEARDOWN_DEFAULT `${APPID}` `${APPNAME}` `${APPEXE}`
 
   !insertmacro APP_UNASSOCIATE ".zip" `${APPID}` `${APPEXE}`
